refactor(phase4b): consolidate anthropic-aisdk into ai/adapters/anthropic
Move the Anthropic AI SDK adapter (client/stream/tools/provider/index) from
api/anthropic-aisdk to the target structure ai/adapters/anthropic (README 4.2),
and leave api/anthropic-aisdk/index.ts as a transparent re-export shim so all
downstream imports keep working unchanged (strangler pattern). Anthropic is
already on @ai-sdk/anthropic, so this is a pure structural move with zero engine
switch and zero runtime behavior change. Internal relative imports rewritten for
the new depth; AbstractBaseProvider now imported from ai/core/BaseProvider
(canonical) instead of the api/baseProvider shim.
